+++ tests/h/util/cors_test.py
@@ -6,7 +6,7 @@ from pyramid.request import Request
 from pyramid.response import Response
 from pyramid.httpexceptions import HTTPBadRequest
 
-from memex.cors import policy, set_cors_headers
+from h.util.cors import policy, set_cors_headers
 
 
 def test_cors_passes_through_non_preflight():
@@ -216,7 +216,7 @@ class TestCorsViewDecorator(object):
 
     @pytest.fixture
     def set_cors_headers(self, patch):
-        return patch('memex.cors.set_cors_headers')
+        return patch('h.util.cors.set_cors_headers')
 
 
 # A tiny WSGI application used for testing the middleware
